# Conflicts:
#	core/src/main/java/org/geysermc/geyser/entity/type/Entity.java
#	core/src/main/java/org/geysermc/geyser/entity/type/TextDisplayEntity.java
#	core/src/main/java/org/geysermc/geyser/session/cache/WorldBorder.java
#	core/src/main/java/org/geysermc/geyser/translator/protocol/bedrock/entity/player/input/BedrockPlayerAuthInputTranslator.java

+++ core/src/main/java/org/geysermc/geyser/registry/loader/ProviderRegistryLoader.java
@@ -213,14 +213,14 @@ public class ProviderRegistryLoader implements RegistryLoader<Map<Class<?>, Prov
         int length = args.length;
         switch (property) {
             case BUNDLE_FULLNESS -> {
-                return new GeyserRangeDispatchPredicate(GeyserRangeDispatchPredicate.GeyserRangeDispatchProperty.BUNDLE_FULLNESS, (int) args[1]);
+                return new GeyserRangeDispatchPredicate(GeyserRangeDispatchPredicate.GeyserRangeDispatchProperty.BUNDLE_FULLNESS, (double) args[1]);
             }
             case DAMAGE -> {
                 // One with, one without normalization
                 if (length == 2) {
                     return new GeyserRangeDispatchPredicate(GeyserRangeDispatchPredicate.GeyserRangeDispatchProperty.DAMAGE, (int) args[1]);
                 } else if (length == 3) {
-                    return new GeyserRangeDispatchPredicate(GeyserRangeDispatchPredicate.GeyserRangeDispatchProperty.DAMAGE, (int) args[1], (boolean) args[2]);
+                    return new GeyserRangeDispatchPredicate(GeyserRangeDispatchPredicate.GeyserRangeDispatchProperty.DAMAGE, (double) args[1], (boolean) args[2]);
                 }
             }
             case COUNT -> {
@@ -228,7 +228,7 @@ public class ProviderRegistryLoader implements RegistryLoader<Map<Class<?>, Prov
                 if (length == 2) {
                     return new GeyserRangeDispatchPredicate(GeyserRangeDispatchPredicate.GeyserRangeDispatchProperty.COUNT, (int) args[1]);
                 } else if (length == 3) {
-                    return new GeyserRangeDispatchPredicate(GeyserRangeDispatchPredicate.GeyserRangeDispatchProperty.COUNT, (int) args[1], (boolean) args[2]);
+                    return new GeyserRangeDispatchPredicate(GeyserRangeDispatchPredicate.GeyserRangeDispatchProperty.COUNT, (double) args[1], (boolean) args[2]);
                 }
             }
             case CUSTOM_MODEL_DATA -> {
